requirements.txt
pandas
google-cloud-bigquery
requests
openpyxl
db-dtypes
pyarrow